requirements.txt
Flask>=3.0
openai>=1.0
pyusb>=1.3